reduce code size, migrate merged to main

+++ lib8812-common/src/main/java/lib8812/common/telemetrymap/FieldConstants.java
@@ -1,11 +1,11 @@
 package lib8812.common.telemetrymap;
 
 public class FieldConstants {
-    public static final int BLOCK_LENGTH_IN = 24;
-    public static final int BLOCK_WIDTH_IN = 24;
-    public static final int BLOCKS_WIDE = 6;
-    public static final int BLOCKS_LONG = 6;
-    public static final int FIELD_LENGTH_IN = BLOCK_LENGTH_IN*BLOCKS_LONG;
-    public static final int FIELD_WIDTH_IN = BLOCK_WIDTH_IN*BLOCKS_WIDE;
+    public static final double BLOCK_LENGTH_IN = 24;
+    public static final double BLOCK_WIDTH_IN = 24;
+    public static final double BLOCKS_WIDE = 6;
+    public static final double BLOCKS_LONG = 6;
+    public static final double FIELD_LENGTH_IN = BLOCK_LENGTH_IN*BLOCKS_LONG;
+    public static final double FIELD_WIDTH_IN = BLOCK_WIDTH_IN*BLOCKS_WIDE;
 
 }
